+++ release-notes.md
@@ -1,14 +1,14 @@
 # Release 1.9.6
 
 ## What's New
-* none
+* [#378](https://github.com/openziti/desktop-edge-win/issues/378) Cleanup old ziti network adapter profiles
 
 ## Other changes:
-* updated installation config
+* updated auto-installation config
 * renewed signing cert
 
 ## Bugs fixed:
-* none
+* set lower metric, if dns server property is set
 
 ## Dependency Updates
 * tsdk is updated to v0.15.6, c-sdk 0.22.5